refactor into spice metrics module
