+++ crates/runtime/src/spice_metrics/metrics_table.rs
@@ -15,9 +15,7 @@ limitations under the License.
 */
 
 use std::any::Any;
-use std::net::SocketAddr;
 use std::sync::Arc;
-use std::time::Duration;
 
 use arrow::array::{Float64Array, StringArray, TimestampSecondArray};
 use arrow::datatypes::{DataType, Field, Schema, TimeUnit};
@@ -30,96 +28,7 @@ use datafusion::logical_expr::Expr;
 use datafusion::physical_plan::memory::MemoryExec;
 use datafusion::physical_plan::ExecutionPlan;
 use snafu::prelude::*;
-use tokio::spawn;
 use tokio::sync::RwLock;
-
-#[derive(Debug, Snafu)]
-pub enum Error {
-    #[snafu(display("Error creating record batch: {source}",))]
-    UnableToCreateRecordBatch { source: arrow::error::ArrowError },
-
-    #[snafu(display("Error queriing prometheus metrics: {source}"))]
-    FailedToQueryPrometheusMetrics { source: reqwest::Error },
-
-    #[snafu(display("Error getting timestamp: {source}"))]
-    UnableToGetTimestamp { source: std::time::SystemTimeError },
-
-    #[snafu(display("Error converting timestamp to seconds"))]
-    UnableToConvertTimestampToSeconds {},
-
-    #[snafu(display("Error parsing prometheus metrics: {source}"))]
-    UnableToParsePrometheusMetrics { source: std::io::Error },
-}
-
-pub struct SpiceMetricsRecorder {
-    socket_addr: Arc<SocketAddr>,
-    table: Arc<MetricsTable>,
-}
-
-impl SpiceMetricsRecorder {
-    pub fn table(&self) -> Arc<MetricsTable> {
-        Arc::clone(&self.table)
-    }
-}
-
-impl SpiceMetricsRecorder {
-    pub fn new(socket_addr: SocketAddr) -> Self {
-        let table = MetricsTable::new();
-
-        Self {
-            socket_addr: Arc::new(socket_addr),
-            table: Arc::new(table),
-        }
-    }
-
-    async fn tick(socket_addr: &SocketAddr, table: &Arc<MetricsTable>) -> Result<(), Error> {
-        let body = reqwest::get(format!("http://{socket_addr}/metrics"))
-            .await
-            .context(FailedToQueryPrometheusMetricsSnafu)?
-            .text()
-            .await
-            .context(FailedToQueryPrometheusMetricsSnafu)?;
-
-        let lines = body.lines().map(|s| Ok(s.to_owned()));
-        let scrape =
-            prometheus_parse::Scrape::parse(lines).context(UnableToParsePrometheusMetricsSnafu)?;
-
-        for sample in scrape.samples {
-            let value: f64 = match sample.value {
-                prometheus_parse::Value::Counter(v)
-                | prometheus_parse::Value::Gauge(v)
-                | prometheus_parse::Value::Untyped(v) => v,
-                prometheus_parse::Value::Histogram(v) => v.into_iter().map(|v| v.count).sum(),
-                prometheus_parse::Value::Summary(v) => v.into_iter().map(|v| v.count).sum(),
-            };
-
-            table
-                .add_record(
-                    sample.timestamp.timestamp(),
-                    &sample.metric,
-                    value,
-                    sample.labels.to_string().as_str(),
-                )
-                .await?;
-        }
-
-        Ok(())
-    }
-
-    pub fn start(&self) {
-        let table = Arc::clone(&self.table);
-        let addr = Arc::clone(&self.socket_addr);
-
-        spawn(async move {
-            loop {
-                if let Err(err) = SpiceMetricsRecorder::tick(&addr, &table).await {
-                    tracing::debug!("{err}");
-                }
-                tokio::time::sleep(Duration::from_secs(30)).await;
-            }
-        });
-    }
-}
 
 pub type PartitionData = Arc<RwLock<Vec<RecordBatch>>>;
 
@@ -169,7 +78,7 @@ impl MetricsTable {
         metric: &str,
         value: f64,
         label: &str,
-    ) -> Result<(), Error> {
+    ) -> Result<(), super::Error> {
         let mut batches = self.batches.write().await;
 
         let batch = RecordBatch::try_new(
@@ -181,7 +90,7 @@ impl MetricsTable {
                 Arc::new(StringArray::from(vec![label])),
             ],
         )
-        .context(UnableToCreateRecordBatchSnafu)?;
+        .context(super::UnableToCreateRecordBatchSnafu)?;
 
         batches.push(Arc::new(RwLock::new(vec![batch])));
